feat:#321 Move test query to test folder
