+++ packages/webapp-libs/webapp-api-client/src/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from './useApiForm';
-export * from './usePaginationQuery';
+export * from './usePaginatedQuery';
