+++ packages/webapp-libs/webapp-api-client/src/tests/factories/pagination.ts
@@ -1,29 +1,6 @@
-import { gql } from '@apollo/client';
-import { TypedDocumentNode } from '@graphql-typed-document-node/core';
 import { PageInfo, composeMockedPaginatedListQueryResult, createDeepFactory, makeId } from '../utils';
-import { PaginationListTestQueryQuery, PaginationListTestQueryQueryVariables } from '@sb/webapp-api-client/graphql';
+import { paginationListTestQuery } from '@sb/webapp-api-client/hooks/usePaginatedQuery/__tests__/usePaginatedQuery.graphql';
 
-
-export const paginationListTestQuery: TypedDocumentNode<
-  PaginationListTestQueryQuery,
-  PaginationListTestQueryQueryVariables
-> = gql(/* GraphQL */ `
-  query paginationListTestQuery($first: Int, $after: String, $last: Int, $before: String) {
-    allNotifications(first: $first, after: $after, last: $last, before: $before) {
-      edges {
-        node {
-          id
-        }
-      }
-      pageInfo {
-        startCursor
-        endCursor
-        hasPreviousPage
-        hasNextPage
-      }
-    }
-  }
-`);
 
 export const paginationTestItemFactory = createDeepFactory<{ id: string; name: string }>(() => ({
   id: makeId(32),
